requirements.txt
# 向量数据库
chromadb>=0.4.0

# 知识图谱数据库
kuzu>=0.4.0

# Token 计数
tiktoken>=0.5.0

# Web 框架（阶段11）
quart>=0.19.0

# JWT 认证（阶段11）
PyJWT>=2.8.0